+++ app/components/ember-tabular-dropdown-limit.js
@@ -1,0 +1,3 @@
+import EmberTabularDropdownLimit from 'ember-tabular/components/ember-tabular-dropdown-limit';
+
+export default EmberTabularDropdownLimit;